// typechain-types/TimeLock.ts
export interface TimeLockInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "deposit"
      | "getLockCount"
      | "getLockInfo"
      | "getTotalLocked"
      | "userLocks"
      | "withdraw"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "Deposited" | "Withdrawn"): EventFragment;

  encodeFunctionData(
    functionFragment: "deposit",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getLockCount",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getLockInfo",
    values: [AddressLike, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getTotalLocked",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "userLocks",
    values: [AddressLike, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "withdraw",
    values: [BigNumberish]
  ): string;

  decodeFunctionResult(functionFragment: "deposit", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "getLockCount",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getLockInfo",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getTotalLocked",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "userLocks", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "withdraw", data: BytesLike): Result;
}

export namespace DepositedEvent {
  export type InputTuple = [
    user: AddressLike,
    amount: BigNumberish,
    unlockTime: BigNumberish,
    lockIndex: BigNumberish
  ];
  export type OutputTuple = [
    user: string,
    amount: bigint,
    unlockTime: bigint,
    lockIndex: bigint
  ];
  export interface OutputObject {
    user: string;
    amount: bigint;
    unlockTime: bigint;
    lockIndex: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace WithdrawnEvent {
  export type InputTuple = [
    user: AddressLike,
    amount: BigNumberish,
    lockIndex: BigNumberish
  ];
  export type OutputTuple = [user: string, amount: bigint, lockIndex: bigint];
  export interface OutputObject {
    user: string;
    amount: bigint;
    lockIndex: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface TimeLock extends BaseContract {
  connect(runner?: ContractRunner | null): TimeLock;
  waitForDeployment(): Promise<this>;

  interface: TimeLockInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  deposit: TypedContractMethod<
    [_lockDuration: BigNumberish],
    [void],
    "payable"
  >;

  getLockCount: TypedContractMethod<[_user: AddressLike], [bigint], "view">;

  getLockInfo: TypedContractMethod<
    [_user: AddressLike, _lockIndex: BigNumberish],
    [
      [bigint, bigint, boolean] & {
        amount: bigint;
        unlockTime: bigint;
        isUnlocked: boolean;
      }
    ],
    "view"
  >;

  getTotalLocked: TypedContractMethod<[_user: AddressLike], [bigint], "view">;

  userLocks: TypedContractMethod<
    [arg0: AddressLike, arg1: BigNumberish],
    [[bigint, bigint] & { amount: bigint; unlockTime: bigint }],
    "view"
  >;

  withdraw: TypedContractMethod<
    [_lockIndex: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "deposit"
  ): TypedContractMethod<[_lockDuration: BigNumberish], [void], "payable">;
  getFunction(
    nameOrSignature: "getLockCount"
  ): TypedContractMethod<[_user: AddressLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "getLockInfo"
  ): TypedContractMethod<
    [_user: AddressLike, _lockIndex: BigNumberish],
    [
      [bigint, bigint, boolean] & {
        amount: bigint;
        unlockTime: bigint;
        isUnlocked: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "getTotalLocked"
  ): TypedContractMethod<[_user: AddressLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "userLocks"
  ): TypedContractMethod<
    [arg0: AddressLike, arg1: BigNumberish],
    [[bigint, bigint] & { amount: bigint; unlockTime: bigint }],
    "view"
  >;
  getFunction(
    nameOrSignature: "withdraw"
  ): TypedContractMethod<[_lockIndex: BigNumberish], [void], "nonpayable">;

  getEvent(
    key: "Deposited"
  ): TypedContractEvent<
    DepositedEvent.InputTuple,
    DepositedEvent.OutputTuple,
    DepositedEvent.OutputObject
  >;
  getEvent(
    key: "Withdrawn"
  ): TypedContractEvent<
    WithdrawnEvent.InputTuple,
    WithdrawnEvent.OutputTuple,
    WithdrawnEvent.OutputObject
  >;

  filters: {
    "Deposited(address,uint256,uint256,uint256)": TypedContractEvent<
      DepositedEvent.InputTuple,
      DepositedEvent.OutputTuple,
      DepositedEvent.OutputObject
    >;
    Deposited: TypedContractEvent<
      DepositedEvent.InputTuple,
      DepositedEvent.OutputTuple,
      DepositedEvent.OutputObject
    >;

    "Withdrawn(address,uint256,uint256)": TypedContractEvent<
      WithdrawnEvent.InputTuple,
      WithdrawnEvent.OutputTuple,
      WithdrawnEvent.OutputObject
    >;
    Withdrawn: TypedContractEvent<
      WithdrawnEvent.InputTuple,
      WithdrawnEvent.OutputTuple,
      WithdrawnEvent.OutputObject
    >;
  };
}